[react] move example components into react-examples
