+++ lib/js/components/dev-layouts/example-react.mjs
@@ -37,7 +37,7 @@ import {
 
 import {
     UIReactTimeControl
-} from '../react-time-control/react-time-control.jsx';
+} from '../react-examples/react-time-control/react-time-control.jsx';
 
 
 import {
@@ -46,11 +46,11 @@ import {
 
 import {
    Counter as ReactCounter
-} from '../react-counter.jsx';
+} from '../react-examples/react-counter.jsx';
 
 import {
    UIReactTimeControl as UIReactTimeControlAlt
-} from '../react-time-control-alternative.jsx';
+} from '../react-examples/react-time-control-alternative.jsx';
 
 
 class ExampleReactController extends _BaseContainerComponent {
