+++ lib/js/components/react-examples/react-time-control-alternative.jsx
@@ -1,6 +1,6 @@
 import React, { useMemo, useCallback } from "react";
 import PropTypes from "prop-types";
-import { useMetamodel } from "./react-integration.jsx";
+import { useMetamodel } from "../react-integration.jsx";
 import "./react-time-control/react-time-control.css";
 
 // Format time as HH:MM:SS.SSS
