--- lib/js/components/react-time-control/index.jsx
@@ -1,6 +1,0 @@
-import { createReactBridge } from "../react-bridge.jsx";
-import { UIReactTimeControl } from "./react-time-control.jsx";
-
-const BridgedUIReactTimeControl = createReactBridge(UIReactTimeControl);
-
-export { BridgedUIReactTimeControl as UIReactTimeControl };
